Split repository.ts: extract roster, policies, docs-ingest failures (4/N)
Fourth slice of the repository.ts carve-up (audit L14), continuing #804, #807
and #809. Three more domains move out behind the same re-export facade:

- repository/roster.ts — server-roster persistence (join/leave transitions and
  the departed-member rows the retention sweep prunes).
- repository/policies.ts — the runtime `policies` rows behind set_policy and
  pause/resume.
- repository/docsIngestFailures.ts — docs-ingest dead-URL tracking (#611).

Pure code motion: all 15 moved symbols verified byte-identical against
origin/main mechanically, and every import site plus tests/repository.test.ts
are untouched. repository.ts is now 6,198 lines, down from 7,137 when the
campaign started.

Screening, per the checklist the earlier PRs established: each candidate was
checked for forward deps (calls out to helpers still in repository.ts), reverse
refs (code in the remaining body referencing something moved — comment-only
mentions don't count), conversation-scoped admin reads (none here, so the 🔒
SQL-scoping invariant is not in play), and tests that slice repository.ts as
source text between `// --- ` banners.

Three candidates were deliberately NOT taken this round:
- Member projects screens clean but tests/tools.test.ts asserts the
  `// --- Member projects` banner exists IN repository.ts, so moving it needs a
  deliberate test change — that belongs in its own PR where the test edit is
  the focus, not buried beside three unrelated moves.
- Shortcut hits and admin audit have reverse code refs (sumShortcutHits,
  recordAdminAction are used by usageStats), so they need explicit imports;
  workable, but kept out to leave this diff a clean move.
- Background job costs is coupled both ways (it calls usageStats).

Note the member-projects banner region in that test now correctly re-resolves
to `// --- Question digest` because roster moved out — which is exactly why
#804 changed it to derive the region end from the next banner instead of
naming a specific following section.
